Get the build to work with Nix
Bazel Java toolchain downloads binaries which depend on FHS, so they do
not work on NixOS. Nix version of Bazel solves the problem by using a
Java toolchain with NONPREBUILT_TOOLCHAIN_CONFIGURATION, which instead
builds the binaries from source. The problem is that the toolchain is only
configured for Java 11, so it doesn't work with this project, since it
uses Java 17.

This is then solved by using rules_nixpkgs which allows configuring a
custom toolchain which sources the JDK from nixpkgs. The toolchain is
used only when --config=nix is specified, which means the build without
nix is still possible as well.

The master branch version of rules_nixpkgs needs to be used, the latest
release 0.10.0 doesn't work.